requirements.txt
dotenv
discord.py
sqlite3
